correción de tipo de datos en clase cache_class

+++ cache_class.cpp
@@ -25,7 +25,7 @@ public:
      * parámetros: indice del bloque (blockIndex), dato a escribir (data), indice al ultimo elemento a copiar (len) 
      */
 
-    void write(int blockIndex, const uint8_t* data, size_t len) {
+    void write(uint8_t blockIndex, const uint8_t* data, uint8_t len) {
         if (blockIndex >= 0 && blockIndex < BLOCKS && len <= BLOCK_SIZE) {
             std::copy(data, data + len, memory[blockIndex].begin()); // data + len es para decir hasta donde debe escribir del array a copiar (data)
             validity[blockIndex] = 1; // Marcar la línea como válida después de escribir
@@ -35,7 +35,7 @@ public:
     /* método para leer 1 linea (bloque) de la cache
      * parámetros: indice del bloque (blockIndex), puntero al arreglo donde se guardará lo leido, indice al ultimo elemento a leer (len) 
      */
-    void read(int blockIndex, uint8_t* outBuffer, size_t len) const {
+    void read(int blockIndex, uint8_t* outBuffer, uint8_t len) const {
         if (blockIndex >= 0 && blockIndex < BLOCKS && len <= BLOCK_SIZE) {
             if (validity[blockIndex] == 1) { // Solo leer si la línea es válida
                 std::copy(memory[blockIndex].begin(), memory[blockIndex].begin() + len, outBuffer);
@@ -48,7 +48,7 @@ public:
     /* método para escribir uno o varios bloques en la cache
      * parámetros: índice del bloque inicial (startBlockIndex), puntero al buffer de entrada (data), número de bloques a escribir (numBlocks)
      */
-    void writeCacheLines(int startBlockIndex, const uint8_t* data, size_t numBlocks) {
+    void writeCacheLines(uint8_t startBlockIndex, const uint8_t* data, uint8_t numBlocks) {
         for (size_t i = 0; i < numBlocks; ++i) {
             write(startBlockIndex + i, data + (i * BLOCK_SIZE), BLOCK_SIZE);
         }
@@ -57,7 +57,7 @@ public:
     /* método para leer uno o varios bloques de la cache
      * parámetros: índice del bloque inicial (startBlockIndex), puntero al buffer de salida (outBuffer), número de bloques a leer (numBlocks)
      */
-    void readCacheLines(int startBlockIndex, uint8_t* outBuffer, size_t numBlocks) const {
+    void readCacheLines(uint8_t startBlockIndex, uint8_t* outBuffer, uint8_t numBlocks) const {
         for (size_t i = 0; i < numBlocks; ++i) {
             read(startBlockIndex + i, outBuffer + (i * BLOCK_SIZE), BLOCK_SIZE);
         }
@@ -66,7 +66,7 @@ public:
     /* método para invalidar una línea de caché
      * parámetros: índice del bloque a invalidar (blockIndex)
      */ 
-    void invalidate(int blockIndex) {
+    void invalidate(uint8_t blockIndex) {
         if (blockIndex >= 0 && blockIndex < BLOCKS) {
             validity[blockIndex] = 0; // Marcar la línea como inválida
         }
@@ -75,7 +75,7 @@ public:
     /* método para verificar si una línea es válida
      * parámetros: índice del bloque a verificar (blockIndex)
      */
-    bool isValid(int blockIndex) const {
+    bool isValid(uint8_t blockIndex) const {
         if (blockIndex >= 0 && blockIndex < BLOCKS) {
             return validity[blockIndex] == 1;
         }
